Affine Expressions

Base.:* — Method

*(lhs, rhs)

Creates an affine mul expression with 'lhs' and 'rhs'.

Base.:+ — Method

+(lhs, rhs)

Creates an affine add expression with 'lhs' and 'rhs'.

Base.:== — Method

==(a, b)

Returns true if the two affine expressions are equal.

Base.cld — Method

cld(lhs, rhs)

Creates an affine ceildiv expression with 'lhs' and 'rhs'.

Base.div — Method

div(lhs, rhs)
÷(lhs, rhs)
fld(lhs, rhs)

Creates an affine floordiv expression with 'lhs' and 'rhs'.

Base.gcd — Method

gcd(affineExpr)

Returns the greatest known integral divisor of this affine expression. The result is always positive.

Base.mod — Method

mod(lhs, rhs)

Creates an affine mod expression with 'lhs' and 'rhs'.

MLIR.IR.AffineDimensionExpr — Method

AffineDimensionExpr(position; context=context)

Creates an affine dimension expression with 'position' in the context.

MLIR.IR.ConstantExpr — Method

ConstantExpr(constant::Int; context=context())

Creates an affine constant expression with 'constant' in the context.

MLIR.IR.SymbolExpr — Method

SymbolExpr(position; context=context())

Creates an affine symbol expression with 'position' in the context.

MLIR.IR.constvalue — Method

constvalue(affineExpr)

Returns the value of the given affine constant expression.

MLIR.IR.context — Method

context(affineExpr)

Gets the context that owns the affine expression.

MLIR.IR.is_pure_affine — Method

is_pure_affine(affineExpr)

Checks whether the given affine expression is a pure affine expression, i.e. mul, floordiv, ceildic, and mod is only allowed w.r.t constants.

MLIR.IR.is_symbolic_or_constant — Method

is_symbolic_or_constant(affineExpr)

Checks whether the given affine expression is made out of only symbols and constants.

MLIR.IR.isadd — Method

isadd(affineExpr)

Checks whether the given affine expression is an add expression.

MLIR.IR.isbinary — Method

isbinary(affineExpr)

Checks whether the given affine expression is binary.

MLIR.IR.isceildiv — Method

isceildiv(affineExpr)

Checks whether the given affine expression is an ceildiv expression.

MLIR.IR.isconstantexpr — Method

isconstantexpr(affineExpr)

Checks whether the given affine expression is a constant expression.

MLIR.IR.isdimexpr — Method

isdimexpr(affineExpr)

Checks whether the given affine expression is a dimension expression.

MLIR.IR.isfloordiv — Method

isfloordiv(affineExpr)

Checks whether the given affine expression is an floordiv expression.

MLIR.IR.isfunctionofdimexpr — Method

isfunctionofdimexpr(affineExpr, position)

Checks whether the given affine expression involves AffineDimExpr 'position'.

MLIR.IR.ismod — Method

ismod(affineExpr)

Checks whether the given affine expression is an mod expression.

MLIR.IR.ismul — Method

ismul(affineExpr)

Checks whether the given affine expression is an mul expression.

MLIR.IR.ismultipleof — Method

ismultipleof(affineExpr, factor)

Checks whether the given affine expression is a multiple of 'factor'.

MLIR.IR.issymbolexpr — Method

issymbolexpr(affineExpr)

Checks whether the given affine expression is a symbol expression.

MLIR.IR.lhs — Method

lhs(affineExpr)

Returns the left hand side affine expression of the given affine binary operation expression.

MLIR.IR.position — Method

position(affineExpr)

Returns the position of the given affine dimension expression, affine symbol expression or ...

MLIR.IR.rhs — Method

rhs(affineExpr)

Returns the right hand side affine expression of the given affine binary operation expression.