Affine Map

MLIR.IR.AffineMap — Method

AffineMap(ndims, nsymbols, affineExprs; context=context())

Creates an affine map with results defined by the given list of affine expressions. The map resulting map also has the requested number of input dimensions and symbols, regardless of them being used in the results.

MLIR.IR.AffineMap — Method

AffineMap(ndims, nsymbols; context=context())

Creates a zero result affine map of the given dimensions and symbols in the context. The affine map is owned by the context.

MLIR.IR.AffineMap — Method

AffineMap(; context=context())

Creates a zero result affine map with no dimensions or symbols in the context. The affine map is owned by the context.

Base.:== — Method

==(a, b)

Checks if two affine maps are equal.

Base.isempty — Method

isempty(affineMap)

Checks whether the given affine map is an empty affine map.

Base.isperm — Method

isperm(affineMap)

Checks whether the given affine map represents a symbol-less permutation map.

Base.ndims — Method

ndims(affineMap)

Returns the number of dimensions of the given affine map.

Base.replace — Method

mlirAffineMapReplace(affineMap, expression => replacement, numResultDims, numResultSyms)

Apply AffineExpr::replace(map) to each of the results and return a new new AffineMap with the new results and the specified number of dims and symbols.

MLIR.IR.ConstantAffineMap — Method

ConstantAffineMap(val; context=context())

Creates a single constant result affine map in the context. The affine map is owned by the context.

MLIR.IR.IdentityAffineMap — Method

IdentityAffineMap(ndims; context=context())

Creates an affine map with 'ndims' identity in the context. The affine map is owned by the context.

MLIR.IR.MinorIdentityAffineMap — Method

MinorIdentityAffineMap(ndims, nresults; context=context())

Creates an identity affine map on the most minor dimensions in the context. The affine map is owned by the context. The function asserts that the number of dimensions is greater or equal to the number of results.

MLIR.IR.PermutationAffineMap — Method

PermutationAffineMap(permutation; context=context())

Creates an affine map with a permutation expression and its size in the context. The permutation expression is a non-empty vector of integers. The elements of the permutation vector must be continuous from 0 and cannot be repeated (i.e. [1,2,0] is a valid permutation. [2,0] or [1,1,2] is an invalid invalid permutation). The affine map is owned by the context.

MLIR.IR.compose — Method

compose(affineExpr, affineMap)

Composes the given map with the given expression.

MLIR.IR.context — Method

context(affineMap)

Gets the context that the given affine map was created with.

MLIR.IR.isidentity — Method

isidentity(affineMap)

Checks whether the given affine map is an identity affine map. The function asserts that the number of dimensions is greater or equal to the number of results.

MLIR.IR.isminoridentity — Method

isminoridentity(affineMap)

Checks whether the given affine map is a minor identity affine map.

MLIR.IR.isprojperm — Method

isprojperm(affineMap)

Checks whether the given affine map represents a subset of a symbol-less permutation map.

MLIR.IR.issingleconstant — Method

issingleconstant(affineMap)

Checks whether the given affine map is a single result constant affine map.

MLIR.IR.majorsubmap — Method

majorsubmap(affineMap, nresults)

Returns the affine map consisting of the most major nresults results. Returns the null AffineMap if the nresults is equal to zero. Returns the affineMap if nresults is greater or equals to number of results of the given affine map.

MLIR.IR.minorsubmap — Method

minorsubmap(affineMap, nresults)

Returns the affine map consisting of the most minor nresults results. Returns the null AffineMap if the nresults is equal to zero. Returns the affineMap if nresults is greater or equals to number of results of the given affine map.

MLIR.IR.ninputs — Method

ninputs(affineMap)

Returns the number of inputs (dimensions + symbols) of the given affine map.

MLIR.IR.nresults — Method

nresults(affineMap)

Returns the number of results of the given affine map.

MLIR.IR.nsymbols — Method

nsymbols(affineMap)

Returns the number of symbols of the given affine map.

MLIR.IR.result — Method

result(affineMap, pos)

Returns the result at the given position.

MLIR.IR.result — Method

result(affineMap)

Returns the constant result of the given affine map. The function asserts that the map has a single constant result.

MLIR.IR.submap — Method

submap(affineMap, positions)

Returns the affine map consisting of the positions subset.

MLIR.IR.@affinemap — Macro

@affinemap (d1, d2, d3, ...)[s1, s2, ...] -> (d0 + d1, ...)

Returns an affine map from the provided Julia expression. On the right hand side are allowed the following function calls:

+, *, ÷, %, fld, cld

The rhs can only contains dimensions and symbols present on the left hand side or integer literals.

julia> using MLIR: IR, AffineUtils

julia> IR.context!(IR.Context()) do
           IR.@affinemap (d1, d2)[s0] -> (d1 + s0, d2 % 10)
       end
MLIR.IR.AffineMap(#= (d0, d1)[s0] -> (d0 + s0, d1 mod 10) =#)