Attribute

Core.Bool — Method

Bool(attr)

Returns the value stored in the given bool attribute.

Core.Float64 — Method

Float64(attr)

Returns the value stored in the given floating point attribute, interpreting the value as double.

Core.Int64 — Method

Int64(attr)

Returns the value stored in the given integer attribute, assuming the value is of signed type and fits into a signed 64-bit integer.

Core.String — Method

String(attr)

Returns the attribute values as a string reference. The data remains live as long as the context in which the attribute lives.

Core.UInt64 — Method

UInt64(attr)

Returns the value stored in the given integer attribute, assuming the value is of unsigned type and fits into an unsigned 64-bit integer.

MLIR.IR.AffineMap — Method

AffineMap(attr)

Returns the affine map wrapped in the given affine map attribute.

MLIR.IR.Attribute — Method

Attribute(str; context=context())

Creates a string attribute in the given context containing the given string.

MLIR.IR.Attribute — Method

Attribute(value; context=context())

Creates a bool attribute in the given context with the given value.

MLIR.IR.Attribute — Method

Attribute(elements; context=context())

Creates a dictionary attribute containing the given list of elements in the provided context.

MLIR.IR.Attribute — Method

Attribute(affineMap)

Creates an affine map attribute wrapping the given map. The attribute belongs to the same context as the affine map.

MLIR.IR.Attribute — Method

Attribute(type, str)

Creates a string attribute in the given context containing the given string. Additionally, the attribute has the given type.

MLIR.IR.Attribute — Method

Attribute(type)

Creates a type attribute wrapping the given type in the same context as the type.

MLIR.IR.Attribute — Method

Attribute(float; context=context(), location=Location(), check=false)

Creates a floating point attribute in the given context with the given double value and double-precision FP semantics. If check=true, emits appropriate diagnostics on illegal arguments.

MLIR.IR.Attribute — Method

Attribute(elements; context=context())

Creates an array element containing the given list of elements in the given context.

MLIR.IR.Attribute — Method

Attribute()

Returns an empty attribute.

MLIR.IR.Attribute — Method

Attribute(int)

Creates an integer attribute of the given type with the given integer value.

MLIR.IR.NamedAttribute — Method

NamedAttribute(name, attr)

Associates an attribute with the name. Takes ownership of neither.

MLIR.IR.Type — Method

Type(attr)

Returns the type stored in the given type attribute.

Base.:== — Method

==(a1, a2)

Checks if two attributes are equal.

Base.fill — Method

fill(attr, shapedType)

Creates a dense elements attribute with the given Shaped type containing a single replicated element (splat).

Base.parse — Method

parse(::Core.Type{Attribute}, str; context=context())

Parses an attribute. The attribute is owned by the context.

Base.reshape — Method

Base.reshape(attr, shapedType)

Creates a dense elements attribute that has the same data as the given dense elements attribute and a different shaped type. The new type must have the same total number of elements.

MLIR.IR.DenseArrayAttribute — Function

DenseArrayAttribute(array; context=context())

Create a dense array attribute with the given elements.

MLIR.IR.DenseElementsAttribute — Method

DenseElementsAttribute(array::AbstractArray{String})

Creates a dense elements attribute with the given shaped type from string elements.

MLIR.IR.DenseElementsAttribute — Method

DenseElementsAttribute(array::AbstractArray)

Creates a dense elements attribute with the given shaped type from elements of a specific type. Expects the element type of the shaped type to match the data element type.

MLIR.IR.DenseElementsAttribute — Method

DenseElementsAttribute(shapedType, elements)

Creates a dense elements attribute with the given Shaped type and elements in the same context as the type.

MLIR.IR.FlatSymbolRefAttribute — Method

FlatSymbolRefAttribute(ctx, symbol)

Creates a flat symbol reference attribute in the given context referencing a symbol identified by the given string.

MLIR.IR.OpaqueAttribute — Method

OpaqueAttribute(dialectNamespace, dataLength, data, type; context=context())

Creates an opaque attribute in the given context associated with the dialect identified by its namespace. The attribute contains opaque byte data of the specified length (data need not be null-terminated).

MLIR.IR.SymbolRefAttribute — Method

SymbolRefAttribute(symbol, references; context=context())

Creates a symbol reference attribute in the given context referencing a symbol identified by the given string inside a list of nested references. Each of the references in the list must not be nested.

MLIR.IR.UnitAttribute — Method

UnitAttribute(; context=context())

Creates a unit attribute in the given context.

MLIR.IR.context — Method

context(attribute)

Gets the context that an attribute was created with.

MLIR.IR.data — Method

data(attr)

Returns the raw data as a string reference. The data remains live as long as the context in which the attribute lives.

MLIR.IR.flatsymbol — Method

flatsymbol(attr)

Returns the referenced symbol as a string reference. The data remains live as long as the context in which the attribute lives.

MLIR.IR.isaffinemap — Method

isaffinemap(attr)

Checks whether the given attribute is an affine map attribute.

MLIR.IR.isarray — Method

isarray(attr)

Checks whether the given attribute is an array attribute.

MLIR.IR.isbool — Method

isbool(attr)

Checks whether the given attribute is a bool attribute.

MLIR.IR.isdensearray — Function

isdensearray(attr, ::Core.Type{T})

Checks whether the given attribute is a dense array attribute.

MLIR.IR.isdenseelements — Method

isdenseelements(attr)

Checks whether the given attribute is a dense elements attribute.

MLIR.IR.isdict — Method

isdict(attr)

Checks whether the given attribute is a dictionary attribute.

MLIR.IR.iselements — Method

iselements(attr)

Checks whether the given attribute is an elements attribute.

MLIR.IR.isflatsymbolref — Method

isflatsymbolref(attr)

Checks whether the given attribute is a flat symbol reference attribute.

MLIR.IR.isfloat — Method

isfloat(attr)

Checks whether the given attribute is a floating point attribute.

MLIR.IR.isinteger — Method

isinteger(attr)

Checks whether the given attribute is an integer attribute.

MLIR.IR.isintegerset — Method

isintegerset(attr)

Checks whether the given attribute is an integer set attribute.

MLIR.IR.isopaque — Method

isopaque(attr)

Checks whether the given attribute is an opaque attribute.

MLIR.IR.isopaqueelements — Method

isopaqueelements(attr)

Checks whether the given attribute is an opaque elements attribute.

MLIR.IR.issparseelements — Method

issparseelements(attr)

Checks whether the given attribute is a sparse elements attribute.

MLIR.IR.issplat — Method

issplat(attr)

Checks whether the given dense elements attribute contains a single replicated value (splat).

MLIR.IR.isstring — Method

isstring(attr)

Checks whether the given attribute is a string attribute.

MLIR.IR.issymbolref — Method

issymbolref(attr)

Checks whether the given attribute is a symbol reference attribute.

MLIR.IR.istype — Method

istype(attr)

Checks whether the given attribute is a type attribute.

MLIR.IR.isunit — Method

isunit(attr)

Checks whether the given attribute is a unit attribute.

MLIR.IR.leafref — Method

leafref(attr)

Returns the string reference to the leaf referenced symbol. The data remains live as long as the context in which the attribute lives.

MLIR.IR.namespace — Method

mlirOpaqueAttrGetDialectNamespace(attr)

Returns the namespace of the dialect with which the given opaque attribute is associated. The namespace string is owned by the context.

MLIR.IR.nnestedrefs — Method

nnestedrefs(attr)

Returns the number of references nested in the given symbol reference attribute.

MLIR.IR.rootref — Method

rootref(attr)

Returns the string reference to the root referenced symbol. The data remains live as long as the context in which the attribute lives.

MLIR.IR.type — Method

type(attribute)

Gets the type of this attribute.

MLIR.IR.typeid — Method

typeid(attribute)

Gets the type id of the attribute.