Block

MLIR.IR.Block — Method

Block(args, locs)

Creates a new empty block with the given argument types and transfers ownership to the caller.

Base.:== — Method

==(block, other)

Checks whether two blocks handles point to the same block. This does not perform deep comparison.

Base.insert! — Method

insert!(block, index, operation)

Takes an operation owned by the caller and inserts it as index to the block. This is an expensive operation that scans the block linearly, prefer insertBefore/After instead.

Base.push! — Method

push!(block, operation)

Takes an operation owned by the caller and appends it to the block.

MLIR.IR.argument — Method

argument(block, i)

Returns i-th argument of the block.

MLIR.IR.first_op — Method

first_op(block)

Returns the first operation in the block or nothing if empty.

MLIR.IR.insert_after! — Method

insert_after!(block, reference, operation)

Takes an operation owned by the caller and inserts it after the (non-owned) reference operation in the given block. If the reference is null, prepends the operation. Otherwise, the reference must belong to the block.

MLIR.IR.insert_before! — Method

insert_before!(block, reference, operation)

Takes an operation owned by the caller and inserts it before the (non-owned) reference operation in the given block. If the reference is null, appends the operation. Otherwise, the reference must belong to the block.

MLIR.IR.nargs — Method

nargs(block)

Returns the number of arguments of the block.

MLIR.IR.next — Method

next(block)

Returns the block immediately following the given block in its parent region or nothing if last.

MLIR.IR.parent_op — Method

parent_op(block)

Returns the closest surrounding operation that contains this block.

MLIR.IR.parent_region — Method

parent_region(block)

Returns the region that contains this block.

MLIR.IR.push_argument! — Method

push_argument!(block, type; location=Location())

Appends an argument of the specified type to the block. Returns the newly added argument.

MLIR.IR.terminator — Method

terminator(block)

Returns the terminator operation in the block or nothing if no terminator.