Integer Set

MLIR.IR.IntegerSet — Method

IntegerSet(ndims, nsymbols, constraints, eqflags; context=context())

Gets or creates a new integer set in the given context. The set is defined by a list of affine constraints, with the given number of input dimensions and symbols, which are treated as either equalities (eqflags is 1) or inequalities (eqflags is 0). Both constraints and eqflags need to be arrays of the same length.

MLIR.IR.IntegerSet — Method

Integerset(ndims, nsymbols; context=context())

Gets or creates a new canonically empty integer set with the give number of dimensions and symbols in the given context.

Base.:== — Method

==(s1, s2)

Checks if two integer set objects are equal. This is a "shallow" comparison of two objects. Only the sets with some small number of constraints are uniqued and compare equal here. Set objects that represent the same integer set with different constraints may be considered non-equal by this check. Set difference followed by an (expensive) emptiness check should be used to check equivalence of the underlying integer sets.

Base.ndims — Method

ndims(set)

Returns the number of dimensions in the given set.

Base.replace — Method

mlirIntegerSetReplaceGet(set, dimReplacements, symbolReplacements, numResultDims, numResultSymbols)

Gets or creates a new integer set in which the values and dimensions of the given set are replaced with the given affine expressions. dimReplacements and symbolReplacements are expected to point to at least as many consecutive expressions as the given set has dimensions and symbols, respectively. The new set will have numResultDims and numResultSymbols dimensions and symbols, respectively.

MLIR.IR.constraint — Method

mlirIntegerSetGetConstraint(set, i)

Returns i-th constraint of the set.

MLIR.IR.context — Method

context(set)

Gets the context in which the given integer set lives.

MLIR.IR.isconstrainteq — Method

mlirIntegerSetIsConstraintEq(set, i)

Returns true of the i-th constraint of the set is an equality constraint, false otherwise.

MLIR.IR.isempty — Method

isempty(set)

Checks whether the given set is a canonical empty set, e.g., the set returned by mlirIntegerSetEmptyGet.

MLIR.IR.nconstraints — Method

nconstraints(set)

Returns the number of constraints (equalities + inequalities) in the given set.

MLIR.IR.nequalities — Method

nequalities(set)

Returns the number of equalities in the given set.

MLIR.IR.ninequalities — Method

ninequalities(set)

Returns the number of inequalities in the given set.

MLIR.IR.ninputs — Method

ninputs(set)

Returns the number of inputs (dimensions + symbols) in the given set.

MLIR.IR.nsymbols — Method

nsymbols(set)

Returns the number of symbols in the given set.