Module

MLIR.IR.Module — Type

Module(location=Location())

Creates a new, empty module and transfers ownership to the caller.

MLIR.IR.Operation — Method

Operation(module)

Views the module as a generic operation.

Base.parse — Method

parse(::Type{Module}, module; context=context())

Parses a module from the string and transfers ownership to the caller.

MLIR.IR.body — Method

body(module)

Gets the body of the module, i.e. the only block it contains.

MLIR.IR.context — Method

context(module)

Gets the context that a module was created with.