Operation

Base.copy — Method

copy(op)

Creates a deep copy of an operation. The operation is not inserted and ownership is transferred to the caller.

MLIR.IR.attr! — Method

attr!(op, name, attr)

Sets an attribute by name, replacing the existing if it exists or adding a new one otherwise.

MLIR.IR.attr — Method

attr(op, name)

Returns an attribute attached to the operation given its name.

MLIR.IR.attr — Method

attr(op, i)

Return i-th attribute of the operation.

MLIR.IR.block — Method

block(op)

Gets the block that owns this operation, returning null if the operation is not owned.

MLIR.IR.context — Method

context(op)

Gets the context this operation is associated with.

MLIR.IR.is_registered — Method

is_registered(name; context=context())

Returns whether the given fully-qualified operation (i.e. 'dialect.operation') is registered with the context. This will return true if the dialect is loaded and the operation is registered within the dialect.

MLIR.IR.location — Method

location(op)

Gets the location of the operation.

MLIR.IR.move_after! — Method

move_after!(op, other)

Moves the given operation immediately after the other operation in its parent block. The given operation may be owned by the caller or by its current block. The other operation must belong to a block. In any case, the ownership is transferred to the block of the other operation.

MLIR.IR.move_before! — Method

move_before!(op, other)

Moves the given operation immediately before the other operation in its parent block. The given operation may be owner by the caller or by its current block. The other operation must belong to a block. In any case, the ownership is transferred to the block of the other operation.

MLIR.IR.name — Method

name(op)

Gets the name of the operation as an identifier.

MLIR.IR.nattrs — Method

nattrs(op)

Returns the number of attributes attached to the operation.

MLIR.IR.noperands — Method

noperands(op)

Returns the number of operands of the operation.

MLIR.IR.nregions — Method

nregions(op)

Returns the number of regions attached to the given operation.

MLIR.IR.nresults — Method

nresults(op)

Returns the number of results of the operation.

MLIR.IR.nsuccessors — Method

nsuccessors(op)

Returns the number of successor blocks of the operation.

MLIR.IR.operand — Function

operand(op, i)

Returns i-th operand of the operation.

MLIR.IR.operand! — Method

operand!(op, i, value)

Sets the i-th operand of the operation.

MLIR.IR.parent_op — Method

parent_op(op)

Gets the operation that owns this operation, returning null if the operation is not owned.

MLIR.IR.region — Method

region(op, i)

Returns i-th region attached to the operation.

MLIR.IR.result — Function

result(op, i)

Returns i-th result of the operation.

MLIR.IR.rmattr! — Method

rmattr!(op, name)

Removes an attribute by name. Returns false if the attribute was not found and true if removed.

MLIR.IR.rmfromparent! — Method

rmfromparent(op)

Removes the given operation from its parent block. The operation is not destroyed. The ownership of the operation is transferred to the caller.

MLIR.IR.successor — Method

successor(op, i)

Returns i-th successor of the operation.

MLIR.IR.typeid — Method

typeid(op)

Gets the type id of the operation. Returns null if the operation does not have a registered operation description.

MLIR.IR.verify — Method

verify(op)

Verify the operation and return true if it passes, false if it fails.