Pass Infrastructure

MLIR.IR.OpPassManager — Method

OpPassManager(opPassManager, operationName)

Nest an OpPassManager under the provided OpPassManager, the nested passmanager will only run on operations matching the provided name. The returned OpPassManager will be destroyed when the parent is destroyed.

MLIR.IR.OpPassManager — Method

OpPassManager(passManager, operationName)

Nest an OpPassManager under the top-level PassManager, the nested passmanager will only run on operations matching the provided name. The returned OpPassManager will be destroyed when the parent is destroyed. To further nest more OpPassManager under the newly returned one, see mlirOpPassManagerNest below.

MLIR.IR.OpPassManager — Method

OpPassManager(passManager)

Cast a top-level PassManager to a generic OpPassManager.

MLIR.IR.PassManager — Method

PassManager(anchorOp; context=context())

Create a new top-level PassManager anchored on anchorOp.

MLIR.IR.PassManager — Method

PassManager(; context=context())

Create a new top-level PassManager.

Base.parse — Method

parse(passManager, pipeline)

Parse a textual MLIR pass pipeline and add it to the provided OpPassManager.

MLIR.IR.add_owned_pass! — Method

add_owned_pass!(opPassManager, pass)

Add a pass and transfer ownership to the provided OpPassManager. If the pass is not a generic operation pass or matching the type of the provided OpPassManager, a new OpPassManager is implicitly nested under the provided OpPassManager.

MLIR.IR.add_owned_pass! — Method

add_owned_pass!(passManager, pass)

Add a pass and transfer ownership to the provided top-level PassManager. If the pass is not a generic operation pass or a ModulePass, a new OpPassManager is implicitly nested under the provided PassManager.

MLIR.IR.add_pipeline! — Method

add_pipeline!(passManager, pipelineElements, callback, userData)

Parse a sequence of textual MLIR pass pipeline elements and add them to the provided OpPassManager. If parsing fails an error message is reported using the provided callback.

MLIR.IR.enable_ir_printing! — Method

enable_ir_printing!(passManager)

Enable mlir-print-ir-after-all.

MLIR.IR.enable_verifier! — Function

enable_verifier!(passManager, enable)

Enable / disable verify-each.

MLIR.IR.run! — Method

run!(passManager, module)

Run the provided passManager on the given module.