Value

Base.:== — Method

==(value1, value2)

Returns 1 if two values are equal, 0 otherwise.

MLIR.IR.block_arg_num — Method

block_arg_num(value)

Returns the position of the value in the argument list of its block.

MLIR.IR.block_owner — Method

block_owner(value)

Returns the block in which this value is defined as an argument. Asserts if the value is not a block argument.

MLIR.IR.is_block_arg — Method

is_block_arg(value)

Returns 1 if the value is a block argument, 0 otherwise.

MLIR.IR.is_op_res — Method

is_op_res(value)

Returns 1 if the value is an operation result, 0 otherwise.

MLIR.IR.op_owner — Method

op_owner(value)

Returns an operation that produced this value as its result. Asserts if the value is not an op result.

MLIR.IR.op_res_num — Method

op_res_num(value)

Returns the position of the value in the list of results of the operation that produced it.

MLIR.IR.type! — Method

set_type!(value, type)

Sets the type of the block argument to the given type.

MLIR.IR.type — Method

type(value)

Returns the type of the value.