amx dialect

MLIR.Dialects.amx.tdpbf16ps — Method

tdpbf16ps

MLIR.Dialects.amx.tdpbssd — Method

tdpbssd

MLIR.Dialects.amx.tdpbsud — Method

tdpbsud

MLIR.Dialects.amx.tdpbusd — Method

tdpbusd

MLIR.Dialects.amx.tdpbuud — Method

tdpbuud

MLIR.Dialects.amx.tile_load — Method

tile_load

Loads a tile from memory defined by a base and indices, with the shape defined by the 2-dim vector type of the result. This is eventually lowered into the "tileloadd" instruction with the corresponding tile configuration.

Example

  %0 = amx.tile_load %arg0[%c0, %c0] : memref<?x?xi8> into vector<16x64xi8>

MLIR.Dialects.amx.tile_mulf — Method

tile_mulf

Multiplies a "m x k" tile with a "k x n" tile and accumulates the results into a "m x n" destination tile. Supports "f32 <- bf16 x bf16" (with pairs of "bf16"). The operation is eventually lowered into the "tdpbf16ps" instruction with the corresponding tile configuration.

Example

  %0 = amx.tile_mulf %a, %b, %c
    : vector<16x32xbf16>, vector<16x32xbf16>, vector<16x16xf32>

MLIR.Dialects.amx.tile_muli — Method

tile_muli

Multiplies a "m x k" tile with a "k x n" tile and accumulates the results into a "m x n" destination tile. Supports all "si32 <- s/ui8 x s/ui8" combinations (4 bytes packed into dwords in the columns of both the source operand tiles; the zero or sign extension is specified with the attributes and default to sign extended). The operation is eventually lowered into one of the "tdpbssd", "tdpbsud", "tdpbusd", or "tdpbuud" instructions with the corresponding tile configuration.

Example

  %0 = amx.tile_muli %a zext, %b zext, %c
    : vector<16x64xi8>, vector<16x64xi8>, vector<16x16xi32>

MLIR.Dialects.amx.tile_store — Method

tile_store

Stores a tile to memory defined by a base and indices, with the shape defined by the 2-dim vector type of the value. This is eventually lowered into the "tilestored" instruction with the corresponding tile configuration.

Example

  amx.tile_store %arg1[%c0, %c0], %0 : memref<?x?xi8>, vector<16x64xi8>

MLIR.Dialects.amx.tile_zero — Method

tile_zero

Zeroes the destination tile, with the shape defined by the 2-dim vector type of the result. This is eventually lowered into the "tilezero" instruction with the corresponding tile configuration.

Example

  %0 = amx.tile_zero : vector<16x16xbf16>

MLIR.Dialects.amx.tileloadd64 — Method

tileloadd64

MLIR.Dialects.amx.tilestored64 — Method

tilestored64

MLIR.Dialects.amx.tilezero — Method

tilezero