arm_sve dialect

MLIR.Dialects.arm_sve.intr_add — Method

intr_add

MLIR.Dialects.arm_sve.intr_fadd — Method

intr_fadd

MLIR.Dialects.arm_sve.intr_fdiv — Method

intr_fdiv

MLIR.Dialects.arm_sve.intr_fmul — Method

intr_fmul

MLIR.Dialects.arm_sve.intr_fsub — Method

intr_fsub

MLIR.Dialects.arm_sve.intr_mul — Method

intr_mul

MLIR.Dialects.arm_sve.intr_sdiv — Method

intr_sdiv

MLIR.Dialects.arm_sve.intr_sdot — Method

intr_sdot

MLIR.Dialects.arm_sve.intr_smmla — Method

intr_smmla

MLIR.Dialects.arm_sve.intr_sub — Method

intr_sub

MLIR.Dialects.arm_sve.intr_udiv — Method

intr_udiv

MLIR.Dialects.arm_sve.intr_udot — Method

intr_udot

MLIR.Dialects.arm_sve.intr_ummla — Method

intr_ummla

MLIR.Dialects.arm_sve.masked_addf — Method

masked_addf

The arm_sve.masked.addf operation takes one scalable vector mask and two scalable vector operands, and perform floating point addition on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_addi — Method

masked_addi

The arm_sve.masked.addi operation takes one scalable vector mask and two scalable vector operands, and perform integer addition on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_divf — Method

masked_divf

The arm_sve.masked.divf operation takes one scalable vector mask and two scalable vector operands, and perform floating point division on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_divi_signed — Method

masked_divi_signed

The arm_sve.masked.divi_signed operation takes one scalable vector mask and two scalable vector operands, and perform integer signed division on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_divi_unsigned — Method

masked_divi_unsigned

The arm_sve.masked.divi_unsigned operation takes one scalable vector mask and two scalable vector operands, and perform integer unsigned division on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_mulf — Method

masked_mulf

The arm_sve.masked.mulf operation takes one scalable vector mask and two scalable vector operands, and perform floating point multiplication on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_muli — Method

masked_muli

The arm_sve.masked.muli operation takes one scalable vector mask and two scalable vector operands, and perform integer multiplication on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_subf — Method

masked_subf

The arm_sve.masked.subf operation takes one scalable vector mask and two scalable vector operands, and perform floating point subtraction on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.masked_subi — Method

masked_subi

The arm_sve.masked.subi operation takes one scalable vector mask and two scalable vector operands, and perform integer subtraction on active lanes. Inactive lanes will keep the value of the first operand.

MLIR.Dialects.arm_sve.sdot — Method

sdot

SDOT: Signed integer addition of dot product.

This function maps to the SDOT instruction, and it takes signless integer operands that the operation interprets as signed. It partitions the second and third vector inputs into groups of four elements. They calculate the dot product of each group (without loss of precision) and then add each result to the overlapping element of the first vector input.

Source: https://developer.arm.com/documentation/100987/0000

MLIR.Dialects.arm_sve.smmla — Method

smmla

SMMLA: Signed integer matrix multiply-accumulate.

This function maps to the SMMLA instruction, and it takes signless integer operands that the operation interprets as signed. It partitions the inputs into 128-bit quadwords, with the first input containing a row-by-row 2×2 matrix of 32-bit integers, the second input containing a row-by-row 2×8 matrix of 8-bit integers, and the third input containing a column-by-column 8×2 matrix of 8-bit integers. For each quadword, they multiply the second input matrix by the third input matrix using natural arithmetic and then add the result to the first input using modular arithmetic.

Source: https://developer.arm.com/documentation/100987/0000

MLIR.Dialects.arm_sve.udot — Method

udot

UDOT: Unsigned integer addition of dot product.

This function maps to the UDOT instruction, and it takes signless integer operands that the operation interprets as unsigned. It partitions the second and third vector inputs into groups of four elements. They calculate the dot product of each group (without loss of precision) and then add each result to the overlapping element of the first vector input.

Source: https://developer.arm.com/documentation/100987/0000

MLIR.Dialects.arm_sve.ummla — Method

ummla

UMMLA: Unsigned integer matrix multiply-accumulate.

This function maps to the UMMLA instruction, and it takes signless integer operands that the operation interprets as unsigned. It partitions the inputs into 128-bit quadwords, with the first input containing a row-by-row 2×2 matrix of 32-bit integers, the second input containing a row-by-row 2×8 matrix of 8-bit integers, and the third input containing a column-by-column 8×2 matrix of 8-bit integers. For each quadword, they multiply the second input matrix by the third input matrix using natural arithmetic and then add the result to the first input using modular arithmetic.

Source: https://developer.arm.com/documentation/100987/0000