async dialect

MLIR.Dialects.async.add_to_group — Method

add_to_group

The async.add_to_group adds an async token or value to the async group. Returns the rank of the added element in the group. This rank is fixed for the group lifetime.

Example

%0 = async.create_group %size : !async.group
%1 = ... : !async.token
%2 = async.add_to_group %1, %0 : !async.token

MLIR.Dialects.async.await — Method

await

The async.await operation waits until the argument becomes ready, and for the async.value arguments it unwraps the underlying value

Example

%0 = ... : !async.token
async.await %0 : !async.token

%1 = ... : !async.value<f32>
%2 = async.await %1 : !async.value<f32>

MLIR.Dialects.async.await_all — Method

await_all

The async.await_all operation waits until all the tokens or values in the group become ready.

Example

%0 = async.create_group %size : !async.group

%1 = ... : !async.token
%2 = async.add_to_group %1, %0 : !async.token

%3 = ... : !async.token
%4 = async.add_to_group %2, %0 : !async.token

async.await_all %0

MLIR.Dialects.async.coro_begin — Method

coro_begin

The async.coro.begin allocates a coroutine frame and returns a handle to the coroutine.

MLIR.Dialects.async.coro_end — Method

coro_end

The async.coro.end marks the point where a coroutine needs to return control back to the caller if it is not an initial invocation of the coroutine. It the start part of the coroutine is is no-op.

MLIR.Dialects.async.coro_free — Method

coro_free

The async.coro.free deallocates the coroutine frame created by the async.coro.begin operation.

MLIR.Dialects.async.coro_id — Method

coro_id

The async.coro.id returns a switched-resume coroutine identifier.

MLIR.Dialects.async.coro_save — Method

coro_save

The async.coro.saves saves the coroutine state.

MLIR.Dialects.async.coro_suspend — Method

coro_suspend

The async.coro.suspend suspends the coroutine and transfers control to the suspend successor. If suspended coroutine later resumed it will transfer control to the resume successor. If it is destroyed it will transfer control to the the cleanup successor.

In switched-resume lowering coroutine can be already in resumed state when suspend operation is called, in this case control will be transferred to the resume successor skipping the suspend successor.

MLIR.Dialects.async.create_group — Method

create_group

The async.create_group allocates an empty async group. Async tokens or values can be added to this group later. The size of the group must be specified at construction time, and await_all operation will first wait until the number of added tokens or values reaches the group size.

Example

%size = ... : index
%group = async.create_group %size : !async.group
...
async.await_all %group

MLIR.Dialects.async.execute — Method

execute

The body region attached to the async.execute operation semantically can be executed concurrently with the successor operation. In the followup example "compute0" can be executed concurrently with "compute1".

The actual concurrency semantics depends on the dialect lowering to the executable format. Fully sequential execution ("compute0" completes before "compute1" starts) is a completely legal execution.

Because concurrent execution is not guaranteed, it is illegal to create an implicit dependency from "compute1" to "compute0" (e.g. via shared global state). All dependencies must be made explicit with async execute arguments (async.token or async.value).

async.execute operation takes async.token dependencies and async.value operands separately, and starts execution of the attached body region only when all tokens and values become ready.

Example

%dependency = ... : !async.token
%value = ... : !async.value<f32>

%token, %results =
  async.execute [%dependency](%value as %unwrapped: !async.value<f32>)
             -> !async.value<!some.type>
  {
    %0 = "compute0"(%unwrapped): (f32) -> !some.type
    async.yield %0 : !some.type
  }

%1 = "compute1"(...) : !some.type

In the example above asynchronous execution starts only after dependency token and value argument become ready. Unwrapped value passed to the attached body region as an %unwrapped value of f32 type.

MLIR.Dialects.async.runtime_add_ref — Method

runtime_add_ref

The async.runtime.add_ref operation adds a reference(s) to async value (token, value or group).

MLIR.Dialects.async.runtime_add_to_group — Method

runtime_add_to_group

The async.runtime.add_to_group adds an async token or value to the async group. Returns the rank of the added element in the group.

MLIR.Dialects.async.runtime_await — Method

runtime_await

The async.runtime.await operation blocks the caller thread until the operand becomes available or error.

MLIR.Dialects.async.runtime_await_and_resume — Method

runtime_await_and_resume

The async.runtime.await_and_resume operation awaits for the operand to become available or error and resumes the coroutine on a thread managed by the runtime.

MLIR.Dialects.async.runtime_create — Method

runtime_create

The async.runtime.create operation creates an async dialect token or value. Tokens and values are created in the non-ready state.

MLIR.Dialects.async.runtime_create_group — Method

runtime_create_group

The async.runtime.create_group operation creates an async dialect group of the given size. Group created in the empty state.

MLIR.Dialects.async.runtime_drop_ref — Method

runtime_drop_ref

The async.runtime.drop_ref operation drops a reference(s) to async value (token, value or group).

MLIR.Dialects.async.runtime_is_error — Method

runtime_is_error

The async.runtime.is_error operation returns true if the token, value or group (any of the async runtime values) is in the error state. It is the caller responsibility to check error state after the call to await or resuming after await_and_resume.

MLIR.Dialects.async.runtime_load — Method

runtime_load

The async.runtime.load operation loads the value from the runtime async.value storage.

MLIR.Dialects.async.runtime_num_worker_threads — Method

runtime_num_worker_threads

The async.runtime.num_worker_threads operation gets the number of threads in the threadpool from the runtime.

MLIR.Dialects.async.runtime_resume — Method

runtime_resume

The async.runtime.resume operation resumes the coroutine on a thread managed by the runtime.

MLIR.Dialects.async.runtime_set_available — Method

runtime_set_available

The async.runtime.set_available operation switches async token or value state to available.

MLIR.Dialects.async.runtime_set_error — Method

runtime_set_error

The async.runtime.set_error operation switches async token or value state to error.

MLIR.Dialects.async.runtime_store — Method

runtime_store

The async.runtime.store operation stores the value into the runtime async.value storage.

MLIR.Dialects.async.yield — Method

yield

The async.yield is a special terminator operation for the block inside async.execute operation.