emitc dialect

MLIR.Dialects.emitc.apply — Method

apply

With the apply operation the operators & (address of) and * (contents of) can be applied to a single operand.

Example

// Custom form of applying the & operator.
%0 = emitc.apply "&"(%arg0) : (i32) -> !emitc.opaque<"int32_t*">

// Generic form of the same operation.
%0 = "emitc.apply"(%arg0) {applicableOperator = "&"}
    : (i32) -> !emitc.opaque<"int32_t*">

MLIR.Dialects.emitc.call — Method

call

The call operation represents a C++ function call. The call allows specifying order of operands and attributes in the call as follows:

integer value of index type refers to an operand;

attribute which will get lowered to constant value in call;

Example

// Custom form defining a call to `foo()`.
%0 = emitc.call "foo" () : () -> i32

// Generic form of the same operation.
%0 = "emitc.call"() {callee = "foo"} : () -> i32

MLIR.Dialects.emitc.cast — Method

cast

The cast operation performs an explicit type conversion and is emitted as a C-style cast expression. It can be applied to integer, float, index and EmitC types.

Example

// Cast from `int32_t` to `float`
%0 = emitc.cast %arg0: i32 to f32

// Cast from `void` to `int32_t` pointer
%1 = emitc.cast %arg1 :
    !emitc.ptr<!emitc.opaque<"void">> to !emitc.ptr<i32>

MLIR.Dialects.emitc.constant — Method

constant

The constant operation produces an SSA value equal to some constant specified by an attribute. This can be used to form simple integer and floating point constants, as well as more exotic things like tensor constants. The constant operation also supports the EmitC opaque attribute and the EmitC opaque type. Since folding is supported, it should not be used with pointers.

Example

// Integer constant
%0 = "emitc.constant"(){value = 42 : i32} : () -> i32

// Constant emitted as `char = CHAR_MIN;`
%1 = "emitc.constant"()
    {value = #emitc.opaque<"CHAR_MIN"> : !emitc.opaque<"char">}
    : () -> !emitc.opaque<"char">

MLIR.Dialects.emitc.include — Method

include

The include operation allows to define a source file inclusion via the #include directive.

Example

// Custom form defining the inclusion of `<myheader>`.
emitc.include <"myheader.h">

// Generic form of the same operation.
"emitc.include" (){include = "myheader.h", is_standard_include} : () -> ()

// Custom form defining the inclusion of `"myheader"`.
emitc.include "myheader.h"

// Generic form of the same operation.
"emitc.include" (){include = "myheader.h"} : () -> ()

MLIR.Dialects.emitc.variable — Method

variable

The variable operation produces an SSA value equal to some value specified by an attribute. This can be used to form simple integer and floating point variables, as well as more exotic things like tensor variables. The variable operation also supports the EmitC opaque attribute and the EmitC opaque type. If further supports the EmitC pointer type, whereas folding is not supported. The variable is emitted as a C/C++ local variable.

Example

// Integer variable
%0 = "emitc.variable"(){value = 42 : i32} : () -> i32

// Variable emitted as `int32_t* = NULL;`
%1 = "emitc.variable"()
    {value = #emitc.opaque<"NULL"> : !emitc.opaque<"int32_t*">}
    : () -> !emitc.opaque<"int32_t*">

Since folding is not supported, it can be used with pointers. As an example, it is valid to create pointers to variable operations by using apply operations and pass these to a call operation.

%0 = "emitc.variable"() {value = 0 : i32} : () -> i32
%1 = "emitc.variable"() {value = 0 : i32} : () -> i32
%2 = emitc.apply "&"(%0) : (i32) -> !emitc.ptr<i32>
%3 = emitc.apply "&"(%1) : (i32) -> !emitc.ptr<i32>
emitc.call "write"(%2, %3) : (!emitc.ptr<i32>, !emitc.ptr<i32>) -> ()