llvm dialect

MLIR.Dialects.llvm.access_group — Method

access_group

Defines an access group metadata that can be attached to any instruction that potentially accesses memory. The access group may be attached to a memory accessing instruction via the llvm.access.group metadata and a branch instruction in the loop latch block via the llvm.loop.parallel_accesses metadata.

See the following link for more details: https://llvm.org/docs/LangRef.html#llvm-access-group-metadata

MLIR.Dialects.llvm.add — Method

add

MLIR.Dialects.llvm.addrspacecast — Method

addrspacecast

MLIR.Dialects.llvm.alias_scope — Method

alias_scope

Defines an alias scope that can be attached to a memory-accessing operation. Such scopes can be used in combination with noalias metadata to indicate that sets of memory-affecting operations in one scope do not alias with memory-affecting operations in another scope.

Example

module { llvm.func @foo(%ptr1 : !llvm.ptr<i32>) { %c0 = llvm.mlir.constant(0 : i32) : i32 %c4 = llvm.mlir.constant(4 : i32) : i32 %1 = llvm.ptrtoint %ptr1 : !llvm.ptr<i32> to i32 %2 = llvm.add %1, %c1 : i32 %ptr2 = llvm.inttoptr %2 : i32 to !llvm.ptr<i32> llvm.store %c0, %ptr1 { aliasscopes = [@metadata::@scope1], llvm.noalias = [@metadata::@scope2] } : !llvm.ptr<i32> llvm.store %c4, %ptr2 { aliasscopes = [@metadata::@scope2], llvm.noalias = [@metadata::@scope1] } : !llvm.ptr<i32> llvm.return }

llvm.metadata @metadata {
  llvm.alias_scope_domain @unused_domain
  llvm.alias_scope_domain @domain { description = "Optional domain description"}
  llvm.alias_scope @scope1 { domain = @domain }
  llvm.alias_scope @scope2 { domain = @domain, description = "Optional scope description" }
  llvm.return
}

}

See the following link for more details: https://llvm.org/docs/LangRef.html#noalias-and-alias-scope-metadata

MLIR.Dialects.llvm.alias_scope_domain — Method

alias_scope_domain

Defines a domain that may be associated with an alias scope.

See the following link for more details: https://llvm.org/docs/LangRef.html#noalias-and-alias-scope-metadata

MLIR.Dialects.llvm.alloca — Method

alloca

MLIR.Dialects.llvm.and — Method

and

MLIR.Dialects.llvm.ashr — Method

ashr

MLIR.Dialects.llvm.atomicrmw — Method

atomicrmw

MLIR.Dialects.llvm.bitcast — Method

bitcast

MLIR.Dialects.llvm.br — Method

br

MLIR.Dialects.llvm.call — Method

call

In LLVM IR, functions may return either 0 or 1 value. LLVM IR dialect implements this behavior by providing a variadic call operation for 0- and 1-result functions. Even though MLIR supports multi-result functions, LLVM IR dialect disallows them.

The call instruction supports both direct and indirect calls. Direct calls start with a function name (@-prefixed) and indirect calls start with an SSA value (%-prefixed). The direct callee, if present, is stored as a function attribute callee. The trailing type of the instruction is always the MLIR function type, which may be different from the indirect callee that has the wrapped LLVM IR function type.

Examples:

// Direct call without arguments and with one result.
%0 = llvm.call @foo() : () -> (f32)

// Direct call with arguments and without a result.
llvm.call @bar(%0) : (f32) -> ()

// Indirect call with an argument and without a result.
llvm.call %1(%0) : (f32) -> ()

MLIR.Dialects.llvm.cmpxchg — Method

cmpxchg

MLIR.Dialects.llvm.cond_br — Method

cond_br

MLIR.Dialects.llvm.extractelement — Method

extractelement

MLIR.Dialects.llvm.extractvalue — Method

extractvalue

MLIR.Dialects.llvm.fadd — Method

fadd

MLIR.Dialects.llvm.fcmp — Method

fcmp

MLIR.Dialects.llvm.fdiv — Method

fdiv

MLIR.Dialects.llvm.fence — Method

fence

MLIR.Dialects.llvm.fmul — Method

fmul

MLIR.Dialects.llvm.fneg — Method

fneg

MLIR.Dialects.llvm.fpext — Method

fpext

MLIR.Dialects.llvm.fptosi — Method

fptosi

MLIR.Dialects.llvm.fptoui — Method

fptoui

MLIR.Dialects.llvm.fptrunc — Method

fptrunc

MLIR.Dialects.llvm.freeze — Method

freeze

MLIR.Dialects.llvm.frem — Method

frem

MLIR.Dialects.llvm.fsub — Method

fsub

MLIR.Dialects.llvm.func — Method

func

MLIR functions are defined by an operation that is not built into the IR itself. The LLVM dialect provides an llvm.func operation to define functions compatible with LLVM IR. These functions have LLVM dialect function type but use MLIR syntax to express it. They are required to have exactly one result type. LLVM function operation is intended to capture additional properties of LLVM functions, such as linkage and calling convention, that may be modeled differently by the built-in MLIR function.

// The type of @bar is !llvm<"i64 (i64)">
llvm.func @bar(%arg0: i64) -> i64 {
  llvm.return %arg0 : i64
}

// Type type of @foo is !llvm<"void (i64)">
// !llvm.void type is omitted
llvm.func @foo(%arg0: i64) {
  llvm.return
}

// A function with `internal` linkage.
llvm.func internal @internal_func() {
  llvm.return
}

MLIR.Dialects.llvm.getelementptr — Method

getelementptr

MLIR.Dialects.llvm.icmp — Method

icmp

MLIR.Dialects.llvm.inline_asm — Method

inline_asm

The InlineAsmOp mirrors the underlying LLVM semantics with a notable exception: the embedded asm_string is not allowed to define or reference any symbol or any global variable: only the operands of the op may be read, written, or referenced. Attempting to define or reference any symbol or any global behavior is considered undefined behavior at this time.

MLIR.Dialects.llvm.insertelement — Method

insertelement

MLIR.Dialects.llvm.insertvalue — Method

insertvalue

MLIR.Dialects.llvm.inttoptr — Method

inttoptr

MLIR.Dialects.llvm.invoke — Method

invoke

MLIR.Dialects.llvm.landingpad — Method

landingpad

MLIR.Dialects.llvm.load — Method

load

MLIR.Dialects.llvm.lshr — Method

lshr

MLIR.Dialects.llvm.metadata — Method

metadata

llvm.metadata op defines one or more metadata nodes.

Example

llvm.metadata @metadata { llvm.accessgroup @group1 llvm.accessgroup @group2 llvm.return }

MLIR.Dialects.llvm.mlir_addressof — Method

mlir_addressof

Creates an SSA value containing a pointer to a global variable or constant defined by llvm.mlir.global. The global value can be defined after its first referenced. If the global value is a constant, storing into it is not allowed.

Examples:

func @foo() {
  // Get the address of a global variable.
  %0 = llvm.mlir.addressof @const : !llvm.ptr<i32>

  // Use it as a regular pointer.
  %1 = llvm.load %0 : !llvm.ptr<i32>

  // Get the address of a function.
  %2 = llvm.mlir.addressof @foo : !llvm.ptr<func<void ()>>

  // The function address can be used for indirect calls.
  llvm.call %2() : () -> ()
}

// Define the global.
llvm.mlir.global @const(42 : i32) : i32

MLIR.Dialects.llvm.mlir_constant — Method

mlir_constant

Unlike LLVM IR, MLIR does not have first-class constant values. Therefore, all constants must be created as SSA values before being used in other operations. llvm.mlir.constant creates such values for scalars and vectors. It has a mandatory value attribute, which may be an integer, floating point attribute; dense or sparse attribute containing integers or floats. The type of the attribute is one of the corresponding MLIR builtin types. It may be omitted for i64 and f64 types that are implied. The operation produces a new SSA value of the specified LLVM IR dialect type. The type of that value must correspond to the attribute type converted to LLVM IR.

Examples:

// Integer constant, internal i32 is mandatory
%0 = llvm.mlir.constant(42 : i32) : i32

// It's okay to omit i64.
%1 = llvm.mlir.constant(42) : i64

// Floating point constant.
%2 = llvm.mlir.constant(42.0 : f32) : f32

// Splat dense vector constant.
%3 = llvm.mlir.constant(dense<1.0> : vector<4xf32>) : vector<4xf32>

MLIR.Dialects.llvm.mlir_global — Method

mlir_global

Since MLIR allows for arbitrary operations to be present at the top level, global variables are defined using the llvm.mlir.global operation. Both global constants and variables can be defined, and the value may also be initialized in both cases.

There are two forms of initialization syntax. Simple constants that can be represented as MLIR attributes can be given in-line:

llvm.mlir.global @variable(32.0 : f32) : f32

This initialization and type syntax is similar to llvm.mlir.constant and may use two types: one for MLIR attribute and another for the LLVM value. These types must be compatible.

More complex constants that cannot be represented as MLIR attributes can be given in an initializer region:

// This global is initialized with the equivalent of:
//   i32* getelementptr (i32* @g2, i32 2)
llvm.mlir.global constant @int_gep() : !llvm.ptr<i32> {
  %0 = llvm.mlir.addressof @g2 : !llvm.ptr<i32>
  %1 = llvm.mlir.constant(2 : i32) : i32
  %2 = llvm.getelementptr %0[%1]
     : (!llvm.ptr<i32>, i32) -> !llvm.ptr<i32>
  // The initializer region must end with `llvm.return`.
  llvm.return %2 : !llvm.ptr<i32>
}

Only one of the initializer attribute or initializer region may be provided.

llvm.mlir.global must appear at top-level of the enclosing module. It uses an @-identifier for its value, which will be uniqued by the module with respect to other @-identifiers in it.

Examples:

// Global values use @-identifiers.
llvm.mlir.global constant @cst(42 : i32) : i32

// Non-constant values must also be initialized.
llvm.mlir.global @variable(32.0 : f32) : f32

// Strings are expected to be of wrapped LLVM i8 array type and do not
// automatically include the trailing zero.
llvm.mlir.global @string("abc") : !llvm.array<3 x i8>

// For strings globals, the trailing type may be omitted.
llvm.mlir.global constant @no_trailing_type("foo bar")

// A complex initializer is constructed with an initializer region.
llvm.mlir.global constant @int_gep() : !llvm.ptr<i32> {
  %0 = llvm.mlir.addressof @g2 : !llvm.ptr<i32>
  %1 = llvm.mlir.constant(2 : i32) : i32
  %2 = llvm.getelementptr %0[%1]
     : (!llvm.ptr<i32>, i32) -> !llvm.ptr<i32>
  llvm.return %2 : !llvm.ptr<i32>
}

Similarly to functions, globals have a linkage attribute. In the custom syntax, this attribute is placed between llvm.mlir.global and the optional constant keyword. If the attribute is omitted, external linkage is assumed by default.

Examples:

// A constant with internal linkage will not participate in linking.
llvm.mlir.global internal constant @cst(42 : i32) : i32

// By default, "external" linkage is assumed and the global participates in
// symbol resolution at link-time.
llvm.mlir.global @glob(0 : f32) : f32

// Alignment is optional
llvm.mlir.global private constant @y(dense<1.0> : tensor<8xf32>) : !llvm.array<8 x f32>

Like global variables in LLVM IR, globals can have an (optional) alignment attribute using keyword alignment. The integer value of the alignment must be a positive integer that is a power of 2.

Examples:

// Alignment is optional
llvm.mlir.global private constant @y(dense<1.0> : tensor<8xf32>) { alignment = 32 : i64 } : !llvm.array<8 x f32>

MLIR.Dialects.llvm.mlir_global_ctors — Method

mlir_global_ctors

Specifies a list of constructor functions and priorities. The functions referenced by this array will be called in ascending order of priority (i.e. lowest first) when the module is loaded. The order of functions with the same priority is not defined. This operation is translated to LLVM's globalctors global variable. The initializer functions are run at load time. The data field present in LLVM's globalctors variable is not modeled here.

Examples:

llvm.mlir.global_ctors {@ctor}

llvm.func @ctor() {
  ...
  llvm.return
}

MLIR.Dialects.llvm.mlir_global_dtors — Method

mlir_global_dtors

Specifies a list of destructor functions and priorities. The functions referenced by this array will be called in descending order of priority (i.e. highest first) when the module is unloaded. The order of functions with the same priority is not defined. This operation is translated to LLVM's globaldtors global variable. The data field present in LLVM's globaldtors variable is not modeled here.

Examples:

llvm.func @dtor() {
  llvm.return
}
llvm.mlir.global_dtors {@dtor}

MLIR.Dialects.llvm.mlir_null — Method

mlir_null

Unlike LLVM IR, MLIR does not have first-class null pointers. They must be explicitly created as SSA values using llvm.mlir.null. This operation has no operands or attributes, and returns a null value of a wrapped LLVM IR pointer type.

Examples:

// Null pointer to i8.
%0 = llvm.mlir.null : !llvm.ptr<i8>

// Null pointer to a function with signature void().
%1 = llvm.mlir.null : !llvm.ptr<func<void ()>>

MLIR.Dialects.llvm.mlir_undef — Method

mlir_undef

Unlike LLVM IR, MLIR does not have first-class undefined values. Such values must be created as SSA values using llvm.mlir.undef. This operation has no operands or attributes. It creates an undefined value of the specified LLVM IR dialect type wrapping an LLVM IR structure type.

Example

// Create a structure with a 32-bit integer followed by a float.
%0 = llvm.mlir.undef : !llvm.struct<(i32, f32)>

MLIR.Dialects.llvm.mul — Method

mul

MLIR.Dialects.llvm.or — Method

or

MLIR.Dialects.llvm.ptrtoint — Method

ptrtoint

MLIR.Dialects.llvm.resume — Method

resume

MLIR.Dialects.llvm.return_ — Method

return_

MLIR.Dialects.llvm.sdiv — Method

sdiv

MLIR.Dialects.llvm.select — Method

select

MLIR.Dialects.llvm.sext — Method

sext

MLIR.Dialects.llvm.shl — Method

shl

MLIR.Dialects.llvm.shufflevector — Method

shufflevector

MLIR.Dialects.llvm.sitofp — Method

sitofp

MLIR.Dialects.llvm.srem — Method

srem

MLIR.Dialects.llvm.store — Method

store

MLIR.Dialects.llvm.sub — Method

sub

MLIR.Dialects.llvm.switch — Method

switch

MLIR.Dialects.llvm.trunc — Method

trunc

MLIR.Dialects.llvm.udiv — Method

udiv

MLIR.Dialects.llvm.uitofp — Method

uitofp

MLIR.Dialects.llvm.unreachable — Method

unreachable

MLIR.Dialects.llvm.urem — Method

urem

MLIR.Dialects.llvm.xor — Method

xor

MLIR.Dialects.llvm.zext — Method

zext