math dialect

MLIR.Dialects.math.abs — Method

abs

The abs operation computes the absolute value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type.

Example

// Scalar absolute value.
%a = math.abs %b : f64

MLIR.Dialects.math.atan — Method

atan

Syntax

operation ::= ssa-id `=` `math.atan` ssa-use `:` type

The atan operation computes the arcus tangent of a given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Arcus tangent of scalar value.
%a = math.atan %b : f64

MLIR.Dialects.math.atan2 — Method

atan2

Syntax

operation ::= ssa-id `=` `math.atan2` ssa-use `,` ssa-use `:` type

The atan2 operation takes two operands and returns one result, all of which must be of the same type. The operands must be of floating point type (i.e., scalar, tensor or vector).

The 2-argument arcus tangent atan2(y, x) returns the angle in the Euclidian plane between the positive x-axis and the ray through the point (x, y). It is a generalization of the 1-argument arcus tangent which returns the angle on the basis of the ratio y/x.

See also https://en.wikipedia.org/wiki/Atan2

Example

// Scalar variant.
%a = math.atan2 %b, %c : f32

MLIR.Dialects.math.ceil — Method

ceil

Syntax

operation ::= ssa-id `=` `math.ceil` ssa-use `:` type

The ceil operation computes the ceiling of a given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar ceiling value.
%a = math.ceil %b : f64

MLIR.Dialects.math.copysign — Method

copysign

Syntax

operation ::= ssa-id `=` `math.copysign` ssa-use `,` ssa-use `:` type

The copysign returns a value with the magnitude of the first operand and the sign of the second operand. It takes two operands and returns one result of the same type. The operands must be of floating point type (i.e., scalar, tensor or vector). It has no standard attributes.

Example

// Scalar copysign value.
%a = math.copysign %b, %c : f64

MLIR.Dialects.math.cos — Method

cos

Syntax

operation ::= ssa-id `=` `math.cos` ssa-use `:` type

The cos operation computes the cosine of a given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar cosine value.
%a = math.cos %b : f64

MLIR.Dialects.math.ctlz — Method

ctlz

The ctlz operation computes the number of leading zeros of an integer value. It operates on scalar, tensor or vector.

Example

// Scalar ctlz function value.
%a = math.ctlz %b : i32

MLIR.Dialects.math.ctpop — Method

ctpop

The ctpop operation computes the number of set bits of an integer value. It operates on scalar, tensor or vector.

Example

// Scalar ctpop function value.
%a = math.ctpop %b : i32

MLIR.Dialects.math.cttz — Method

cttz

The cttz operation computes the number of trailing zeros of an integer value. It operates on scalar, tensor or vector.

Example

// Scalar cttz function value.
%a = math.cttz %b : i32

MLIR.Dialects.math.erf — Method

erf

Syntax

operation ::= ssa-id `=` `math.erf` ssa-use `:` type

The erf operation computes the error function. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar error function value.
%a = math.erf %b : f64

MLIR.Dialects.math.exp — Method

exp

Syntax

operation ::= ssa-id `=` `math.exp` ssa-use `:` type

The exp operation takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar natural exponential.
%a = math.exp %b : f64

MLIR.Dialects.math.exp2 — Method

exp2

Syntax

operation ::= ssa-id `=` `math.exp2` ssa-use `:` type

The exp operation takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar natural exponential.
%a = math.exp2 %b : f64

MLIR.Dialects.math.expm1 — Method

expm1

Syntax

operation ::= ssa-id `=` `math.expm1` ssa-use `:` type

expm1(x) := exp(x) - 1

The expm1 operation takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar natural exponential minus 1.
%a = math.expm1 %b : f64

MLIR.Dialects.math.floor — Method

floor

Syntax

operation ::= ssa-id `=` `math.floor` ssa-use `:` type

The floor operation computes the floor of a given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar floor value.
%a = math.floor %b : f64 

MLIR.Dialects.math.fma — Method

fma

Syntax

operation ::= ssa-id `=` `math.fma` ssa-use `,` ssa-use `,` ssa-use `:` type

The fma operation takes three operands and returns one result, each of these is required to be the same type. Operands must be of floating point type (i.e., scalar, tensor or vector).

Example

// Scalar fused multiply-add: d = a*b + c
%d = math.fma %a, %b, %c : f64

The semantics of the operation correspond to those of the llvm.fma intrinsic. In the particular case of lowering to LLVM, this is guaranteed to lower to the llvm.fma.* intrinsic.

MLIR.Dialects.math.log — Method

log

Computes the base-e logarithm of the given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type.

Example

// Scalar log operation.
%y = math.log %x : f64

MLIR.Dialects.math.log10 — Method

log10

Computes the base-10 logarithm of the given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type.

Example

// Scalar log10 operation.
%y = math.log10 %x : f64

MLIR.Dialects.math.log1p — Method

log1p

Computes the base-e logarithm of one plus the given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type.

log1p(x) := log(1 + x)

Example

// Scalar log1p operation.
%y = math.log1p %x : f64

MLIR.Dialects.math.log2 — Method

log2

Computes the base-2 logarithm of the given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type.

Example

// Scalar log2 operation.
%y = math.log2 %x : f64

MLIR.Dialects.math.powf — Method

powf

Syntax

operation ::= ssa-id `=` `math.powf` ssa-use `,` ssa-use `:` type

The powf operation takes two operands of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. Operands must have the same type.

Example

// Scalar exponentiation.
%a = math.powf %b, %c : f64

MLIR.Dialects.math.round — Method

round

Syntax

operation ::= ssa-id `=` `math.round` ssa-use `:` type

The round operation returns the operand rounded to the nearest integer value in floating-point format. It takes one operand of floating point type (i.e., scalar, tensor or vector) and produces one result of the same type. The operation rounds the argument to the nearest integer value in floating-point format, rounding halfway cases away from zero, regardless of the current rounding direction.

Example

// Scalar round operation.
%a = math.round %b : f64

MLIR.Dialects.math.rsqrt — Method

rsqrt

The rsqrt operation computes the reciprocal of the square root. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar reciprocal square root value.
%a = math.rsqrt %b : f64

MLIR.Dialects.math.sin — Method

sin

Syntax

operation ::= ssa-id `=` `math.sin` ssa-use `:` type

The sin operation computes the sine of a given value. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar sine value.
%a = math.sin %b : f64

MLIR.Dialects.math.sqrt — Method

sqrt

The sqrt operation computes the square root. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar square root value.
%a = math.sqrt %b : f64

MLIR.Dialects.math.tan — Method

tan

The tan operation computes the tangent. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar tangent value.
%a = math.tan %b : f64

MLIR.Dialects.math.tanh — Method

tanh

The tanh operation computes the hyperbolic tangent. It takes one operand of floating point type (i.e., scalar, tensor or vector) and returns one result of the same type. It has no standard attributes.

Example

// Scalar hyperbolic tangent value.
%a = math.tanh %b : f64