acc dialect

MLIR.Dialects.acc.data — Function

data

The "acc.data" operation represents a data construct. It defines vars to be allocated in the current device memory for the duration of the region, whether data should be copied from local memory to the current device memory upon region entry , and copied from device memory to local memory upon region exit.

Example

acc.data present(%a: memref<10x10xf32>, %b: memref<10x10xf32>,
    %c: memref<10xf32>, %d: memref<10xf32>) {
  // data region
}

MLIR.Dialects.acc.enter_data — Function

enter_data

The "acc.enter_data" operation represents the OpenACC enter data directive.

Example

acc.enter_data create(%d1 : memref<10xf32>) attributes {async}

MLIR.Dialects.acc.exit_data — Function

exit_data

The "acc.exit_data" operation represents the OpenACC exit data directive.

Example

acc.exit_data delete(%d1 : memref<10xf32>) attributes {async}

MLIR.Dialects.acc.init — Function

init

The "acc.init" operation represents the OpenACC init executable directive.

Example

acc.init
acc.init device_num(%dev1 : i32)

MLIR.Dialects.acc.loop — Function

loop

The "acc.loop" operation represents the OpenACC loop construct.

Example

acc.loop gang vector {
  scf.for %arg3 = %c0 to %c10 step %c1 {
    scf.for %arg4 = %c0 to %c10 step %c1 {
      scf.for %arg5 = %c0 to %c10 step %c1 {
        // ... body
      }
    }
  }
  acc.yield
} attributes { collapse = 3 }

MLIR.Dialects.acc.parallel — Function

parallel

The "acc.parallel" operation represents a parallel construct block. It has one region to be executed in parallel on the current device.

Example

acc.parallel num_gangs(%c10) num_workers(%c10)
    private(%c : memref<10xf32>) {
  // parallel region
}

MLIR.Dialects.acc.shutdown — Function

shutdown

The "acc.shutdown" operation represents the OpenACC shutdown executable directive.

Example

acc.shutdown
acc.shutdown device_num(%dev1 : i32)

MLIR.Dialects.acc.terminator — Method

terminator

A terminator operation for regions that appear in the body of OpenACC operation. Generic OpenACC construct regions are not expected to return any value so the terminator takes no operands. The terminator op returns control to the enclosing op.

MLIR.Dialects.acc.update — Function

update

The "acc.udpate" operation represents the OpenACC update executable directive. As host and self clauses are synonyms, any operands for host and self are add to hostOperands.

Example

acc.update device(%d1 : memref<10xf32>) attributes {async}

MLIR.Dialects.acc.wait — Function

wait

The "acc.wait" operation represents the OpenACC wait executable directive.

Example

acc.wait(%value1: index)
acc.wait() async(%async1: i32)

MLIR.Dialects.acc.yield — Method

yield

acc.yield is a special terminator operation for block inside regions in acc ops (parallel and loop). It returns values to the immediately enclosing acc op.