quant dialect

MLIR.Dialects.quant.const_fake_quant — Method

const_fake_quant

Given a const min, max, numbits and narrowrange attribute, applies the same uniform quantization simulation as is done by the TensorFlow fakequantwithminmax_args op. See the fakeQuantAttrsToType() utility method and the quant-convert-simulated-quantization pass for further details.

MLIR.Dialects.quant.const_fake_quant_per_axis — Method

const_fake_quant_per_axis

Given a const min, max, numbits and narrowrange attribute, applies the same per axis uniform quantization simulation as is done by the TensorFlow fakequantwithminmaxvarsper_channel op. See the fakeQuantAttrsToType() utility method and the quant-convert-simulated-quantization pass for further details.

MLIR.Dialects.quant.coupled_ref — Method

coupled_ref

Ordinarily, relationships between ops for the purposes of determining compatible quantized types is explicit based on the use-def chain. However, in some situations, a use may be separated from its def by arbitrary external connections. In such a case, during analysis, all coupled_ref nodes in a module which share a coupledKey will be considered to be directly connected as via an identity op for the purpose of type inference.

MLIR.Dialects.quant.dcast — Method

dcast

MLIR.Dialects.quant.qcast — Method

qcast

MLIR.Dialects.quant.region — Method

region

MLIR.Dialects.quant.return_ — Method

return_

MLIR.Dialects.quant.scast — Method

scast

MLIR.Dialects.quant.stats — Method

stats

Associates statistics about the runtime ranges of values observed for evaluations of this node.

Statistics about the entire type are reported in the 'layerStats' attribute and those for each axis, in the (optional) axisStats attribute. The interpretation of each is determined by the last dimension of its shape. Currently, only dim=2 is supported, which is interpreted as [min, max].

layerStats must be a rank 1 tensor: [2] axisStats must be a rank 2 tensor: [N, 2], where N=the slice size splitted by the axis dimension. For example:

<?x?x3x2>, axis=3 => N=2
<?x?x3x2>, axis=2 => N=6

MLIR.Dialects.quant.stats_ref — Method

stats_ref

This op acts as an identity that, when encountered at runtime, should result in statistics being collected about about the value of its operand/result. Such statistics will be stored with the provided key, allowing this node to later be converted to a 'stats' op if statistics with that key have been encountered.