tosa dialect

MLIR.Dialects.tosa.abs — Method

abs

Elementwise absolute value operation

MLIR.Dialects.tosa.add — Method

add

Elementwise addition of input1 and input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.apply_scale — Method

apply_scale

Applies rescaling for fixed point values. This behavior is replicated in multiple quantized operations (mul, convolution, rescale, matmul, pooling).

The commonplace implementation is to use i64 operations to avoid integer overflow with target specific implementations can use native operations to avoid wider than necessary types.

MLIR.Dialects.tosa.argmax — Method

argmax

This returns the index with the largest value across the given axis of the input tensor.

MLIR.Dialects.tosa.arithmetic_right_shift — Method

arithmetic_right_shift

Elementwise arithmetic right shift of input1 by the amount specified in input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.avg_pool2d — Method

avg_pool2d

This performs an average pooling over the given input tensor. A sliding window of size given by <kernel size> is passed over the input tensor, with the mean value being placed in the output tensor.

MLIR.Dialects.tosa.bitwise_and — Method

bitwise_and

Elementwise bitwise AND of input1 and input2. Axis of size 1 will be broadcast as necessary.

MLIR.Dialects.tosa.bitwise_not — Method

bitwise_not

Elementwise bitwise NOT of input tensor.

MLIR.Dialects.tosa.bitwise_or — Method

bitwise_or

Elementwise bitwise OR of input1 and input2. Axis of size 1 will be broadcast as necessary.

MLIR.Dialects.tosa.bitwise_xor — Method

bitwise_xor

Elementwise bitwise XOR of input1 and input2. Axis of size 1 will be broadcast as necessary.

MLIR.Dialects.tosa.cast — Method

cast

Performs a set of permissible cast operations Mode Input Output –––––––––––––––––––- signed 8 to bool int8 Boolean signed 16 to bool int16 Boolean signed 32 to bool int32 Boolean bool to 8 Boolean int8 bool to 16 Boolean int16 bool to 32 Boolean int32 signed 8 to signed 16 int8 int16 signed 8 to signed 32 int8 int32 signed 16 to signed 8 int16 int8 signed 16 to signed 32 int16 int32 signed 32 to signed 8 int32 int8 signed 32 to signed 16 int32 int16 float to signed 8 float int8 float to signed 16 float int16 signed 8 to float int8 float signed 16 to float int16 float

MLIR.Dialects.tosa.ceil — Method

ceil

Elementwise ceiling operation

MLIR.Dialects.tosa.clamp — Method

clamp

Clamp to an arbitrary minimum and maximum value. Maximum and minimum values are specified as values in the range of the input type. No zero point subtraction is done to the values, thus to clamp to the zero point value, the zero point itself should be supplied as the minimum value.

MLIR.Dialects.tosa.clz — Method

clz

Elementwise count leading zeros operation

MLIR.Dialects.tosa.concat — Method

concat

Concatenate a variadic amount of tensors along a given axis. No data conversion happens during a concat operation.

MLIR.Dialects.tosa.cond_if — Method

cond_if

Evaluates a Boolean condition and then takes one of two distinct execution paths. This implements the semantic If-then-else structure.

MLIR.Dialects.tosa.const_ — Method

const_

A node containing constant data for use as the input to an operation. May hold data in any of the supported data formats.

MLIR.Dialects.tosa.conv2d — Method

conv2d

Performs a 2D convolution over the given tensor input, using the weight tensor.

MLIR.Dialects.tosa.conv3d — Method

conv3d

Performs a 3D convolution over the given input tensor.

MLIR.Dialects.tosa.custom — Method

custom

Hardware implementing TOSA may choose to add additional custom operators that are not expressed in the existing TOSA operations. These operators are not expected to be portable across TOSA implementations. The input and output signatures must be expressed in the corresponding TOSA node.

MLIR.Dialects.tosa.depthwise_conv2d — Method

depthwise_conv2d

Performs 2D convolutions separately over each channel of the given tensor input, using the weight tensor.

MLIR.Dialects.tosa.div — Method

div

Elementwise integer divide operator of input1 by input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.equal — Method

equal

Elementwise comparison operation

MLIR.Dialects.tosa.exp — Method

exp

Elementwise e to the x operation

MLIR.Dialects.tosa.floor — Method

floor

Elementwise floor operation

MLIR.Dialects.tosa.fully_connected — Method

fully_connected

Performs a fully connected network.

MLIR.Dialects.tosa.gather — Method

gather

Generate a tensor for which each element in the output is a slice of the values tensor based on the value of indices.

MLIR.Dialects.tosa.greater — Method

greater

Elementwise greater than comparison operation

MLIR.Dialects.tosa.greater_equal — Method

greater_equal

Elementwise comparison operation

MLIR.Dialects.tosa.identity — Method

identity

Returns a tensor with the same shape, size, type and content as the input.

MLIR.Dialects.tosa.log — Method

log

Elementwise natural logarithm operation

MLIR.Dialects.tosa.logical_and — Method

logical_and

Elementwise logical AND of input1 and input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.logical_left_shift — Method

logical_left_shift

Elementwise left shift of input1 and input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.logical_not — Method

logical_not

Elementwise logical NOT of input.

MLIR.Dialects.tosa.logical_or — Method

logical_or

Elementwise logical OR of input1 and input2. Axis of size 1 will be broadcast as necessary.

MLIR.Dialects.tosa.logical_right_shift — Method

logical_right_shift

Elementwise logical right shift of input1 by the amount specified in input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.logical_xor — Method

logical_xor

Elementwise logical XOR of input1 and input2. Axis of size 1 will be broadcast as necessary.

MLIR.Dialects.tosa.matmul — Method

matmul

Performs a two dimensional matrix multiplication. This allows both inputs to be activations, rather than reserving weights as an attribute in the FULLY_CONNECTED operator.

MLIR.Dialects.tosa.max_pool2d — Method

max_pool2d

This performs a max pooling over the given input tensor. A sliding window of size given by <kernel size> is passed over the input tensor, with the maximum value being placed in the output tensor.

MLIR.Dialects.tosa.maximum — Method

maximum

Elementwise max of input1 and input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.minimum — Method

minimum

Elementwise minimum of input1 and input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.mul — Method

mul

Elementwise multiplication (Hadamard product) of input1 and input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.negate — Method

negate

Elementwise negation operation

MLIR.Dialects.tosa.pad — Function

pad

Pads a tensor along borders of each dimension with pad_value.

MLIR.Dialects.tosa.pow — Method

pow

Elementwise input1 raised to the power of input2. Axis of size 1 will be broadcast, as necessary.

MLIR.Dialects.tosa.reciprocal — Method

reciprocal

Elementwise reciprocal operation. For integer operation, a TABLE should be used with the appropriate ranges.

MLIR.Dialects.tosa.reduce_all — Method

reduce_all

Reduce a tensor along the given axis with a logical AND operation

MLIR.Dialects.tosa.reduce_any — Method

reduce_any

Reduce a tensor along the given axis with a logical OR operation

MLIR.Dialects.tosa.reduce_max — Method

reduce_max

Reduce a tensor along the given axis with a maximum operation

MLIR.Dialects.tosa.reduce_min — Method

reduce_min

Reduce a tensor along the given axis with a minimum operation

MLIR.Dialects.tosa.reduce_prod — Method

reduce_prod

Reduce a tensor along the given axis by computing the product of the axis.

MLIR.Dialects.tosa.reduce_sum — Method

reduce_sum

Reduce a tensor along the given axis by computing the sum of the axis.

MLIR.Dialects.tosa.reluN — Method

reluN

ReLU with a scalar maximum value.

MLIR.Dialects.tosa.rescale — Method

rescale

Rescale quantized values into a new domain. Supported rescalings are: Mode Input Output signed 8 to 8 int8 int8 signed 8 to 16 int8 int16 signed 8 to 32 int8 int32 signed 16 to 8 int16 int8 signed 16 to 16 int16 int16 signed 16 to 32 int16 int32 signed 32 to 8 int32 int8 signed 32 to 16 int32 int16 signed 32 to 32 int32 int32 signed 48 to 8 int48 int8 signed 48 to 16 int48 int16 signed 48 to 32 int48 int32 unsigned 8 to signed 8 uint8 int8 signed 8 to unsigned 8 int8 uint8

MLIR.Dialects.tosa.reshape — Method

reshape

Returns a tensor with the same type/values as the input, with a new shape specified by the shape argument. Reshape may operate on tensors of any rank. No data conversion happens during a reshape operation.

MLIR.Dialects.tosa.resize — Method

resize

Resizes a tensor. Resize is only allowed in the H and W dimensions. In expected use, stridey is approximately (IH<<shift)/OH and stridex is approximately (IW<<shift)/OW. OH and OW are also supplied as inputs since there may be off by one errors if calculating OH and OW from the strides.

MLIR.Dialects.tosa.reverse — Method

reverse

Returns a tensor with the same type/values as the input, with the data reversed along the given axis. No data conversion happens during a reverse operation.

MLIR.Dialects.tosa.rsqrt — Method

rsqrt

Elementwise reciprocal square root operation. For integer operation, a TABLE should be used with the appropriate ranges.

MLIR.Dialects.tosa.scatter — Method

scatter

The valuesout tensor is set to the valuesin tensor with data modified as follows: data from the input tensor is inserted at the positions specified by the indices tensor.

MLIR.Dialects.tosa.select — Method

select

Elementwise select of the output based on a condition.

MLIR.Dialects.tosa.sigmoid — Method

sigmoid

Sigmoid function: output = 1 / (1 + exp(-input)) For quantized integer data types, the TABLE operator should be used instead with the following definition. The sigmoid table has 513 entries each of 16-bit precision and covering the input range -16.0 to +16.0 in steps of 1/16.

MLIR.Dialects.tosa.slice — Method

slice

Extracts a slice of the input1 on the given axis, beginning at the start coordinates, and extending for size elements in each direction. No data conversion happens during a slice operation.

MLIR.Dialects.tosa.sub — Method

sub

Elementwise subtraction of input1 and input2. Axis of size 1 will be broadcast as necessary.

MLIR.Dialects.tosa.table — Method

table

Interpolated table lookup operation. Input values are scaled to create a fixed-point 9.7 value. The high 9 bits are used to index into the table. The fractional bits are used to interpolate based on the looked up value and the index+1 value in the table. The TABLE operator then returns a 16.7 interpolated value. Note that there must be 513 values to handle the full range of inputs.

The TABLE operator is expected to be used as follows:

A RESCALE node is expected before the TABLE operator to scale the input to a full int16_t range for the table lookup

If an int16_t result is required then follow the TABLE operator with a RESCALE with a right shift of 7

If an int8_t result is required then follow the TABLE operator with a RESCALE with a right shift of 15

MLIR.Dialects.tosa.tanh — Method

tanh

Parameterized hyperbolic tangent. For quantized integer data types, the TABLE operator should be used instead with the following definition. The tanh_table has 513 entries each of 16-bit precision and covering the input range -8.0 to +8.0 in steps of 1/32.

MLIR.Dialects.tosa.tile — Method

tile

Replicates input 0 multiplies times along each dimension.

MLIR.Dialects.tosa.transpose — Method

transpose

Permutes the dimensions based on perm.

MLIR.Dialects.tosa.transpose_conv2d — Method

transpose_conv2d

Performs a 2D transposed convolution over the given tensor input, using the weights tensor.

MLIR.Dialects.tosa.while_loop — Method

while_loop

Generates and evaluates a Bool condition and either executes a loop body or exits to another control point. This action is performed repeatedly after updating and re-evaluating the Boolean condition every iteration. This implements the semantic foreach or while iterative loop structure.

MLIR.Dialects.tosa.yield — Method

yield

return operation within the conditional and body of structured control flow. Operation takes variadic operands but produces no results of its own.