x86vector dialect

MLIR.Dialects.x86vector.avx512_intr_mask_compress — Method

avx512_intr_mask_compress

MLIR.Dialects.x86vector.avx512_intr_mask_rndscale_pd_512 — Method

avx512_intr_mask_rndscale_pd_512

MLIR.Dialects.x86vector.avx512_intr_mask_rndscale_ps_512 — Method

avx512_intr_mask_rndscale_ps_512

MLIR.Dialects.x86vector.avx512_intr_mask_scalef_pd_512 — Method

avx512_intr_mask_scalef_pd_512

MLIR.Dialects.x86vector.avx512_intr_mask_scalef_ps_512 — Method

avx512_intr_mask_scalef_ps_512

MLIR.Dialects.x86vector.avx512_intr_vp2intersect_d_512 — Method

avx512_intr_vp2intersect_d_512

MLIR.Dialects.x86vector.avx512_intr_vp2intersect_q_512 — Method

avx512_intr_vp2intersect_q_512

MLIR.Dialects.x86vector.avx512_mask_compress — Function

avx512_mask_compress

The mask.compress op is an AVX512 specific op that can lower to the llvm.mask.compress instruction. Instead of src, a constant vector vector attribute constant_src may be specified. If neither src nor constant_src is specified, the remaining elements in the result vector are set to zero.

From the Intel Intrinsics Guide:

Contiguously store the active integer/floating-point elements in a (those with their respective bit set in writemask k) to dst, and pass through the remaining elements from src.

MLIR.Dialects.x86vector.avx512_mask_rndscale — Method

avx512_mask_rndscale

The mask.rndscale op is an AVX512 specific op that can lower to the proper LLVMAVX512 operation: llvm.mask.rndscale.ps.512 or llvm.mask.rndscale.pd.512 instruction depending on the type of vectors it is applied to.

From the Intel Intrinsics Guide:

Round packed floating-point elements in a to the number of fraction bits specified by imm, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

MLIR.Dialects.x86vector.avx512_mask_scalef — Method

avx512_mask_scalef

The mask.scalef op is an AVX512 specific op that can lower to the proper LLVMAVX512 operation: llvm.mask.scalef.ps.512 or llvm.mask.scalef.pd.512 depending on the type of MLIR vectors it is applied to.

From the Intel Intrinsics Guide:

Scale the packed floating-point elements in a using values from b, and store the results in dst using writemask k (elements are copied from src when the corresponding mask bit is not set).

MLIR.Dialects.x86vector.avx512_vp2intersect — Method

avx512_vp2intersect

The vp2intersect op is an AVX512 specific op that can lower to the proper LLVMAVX512 operation: llvm.vp2intersect.d.512 or llvm.vp2intersect.q.512 depending on the type of MLIR vectors it is applied to.

From the Intel Intrinsics Guide:

Compute intersection of packed integer vectors a and b, and store indication of match in the corresponding bit of two mask registers specified by k1 and k2. A match in corresponding elements of a and b is indicated by a set bit in the corresponding bit of the mask registers.

MLIR.Dialects.x86vector.avx_intr_dot — Method

avx_intr_dot

Computes the 4-way dot products of the lower and higher parts of the source vectors and broadcasts the two results to the lower and higher elements of the destination vector, respectively. Adding one element of the lower part to one element of the higher part in the destination vector yields the full dot product of the two source vectors.

Example

%0 = x86vector.avx.intr.dot %a, %b : vector<8xf32>
%1 = vector.extractelement %0[%i0 : i32]: vector<8xf32>
%2 = vector.extractelement %0[%i4 : i32]: vector<8xf32>
%d = arith.addf %1, %2 : f32

MLIR.Dialects.x86vector.avx_intr_dp_ps_256 — Method

avx_intr_dp_ps_256

MLIR.Dialects.x86vector.avx_intr_rsqrt_ps_256 — Method

avx_intr_rsqrt_ps_256

MLIR.Dialects.x86vector.avx_rsqrt — Method

avx_rsqrt