Type

MLIR.IR.Type — Method

Type(T::Core.Type{<:Integer}; context=context()

Creates a signless integer type of the given bitwidth in the context. The type is owned by the context.

MLIR.IR.Type — Method

Type(T::Core.Type{<:Signed}; context=context()

Creates a signed integer type of the given bitwidth in the context. The type is owned by the context.

MLIR.IR.Type — Method

Type(T::Core.Type{<:Unsigned}; context=context()

Creates an unsigned integer type of the given bitwidth in the context. The type is owned by the context.

MLIR.IR.Type — Method

Type(T::Core.Type{Bool}; context=context()

Creates a 1-bit signless integer type in the context. The type is owned by the context.

MLIR.IR.Type — Method

Type(::Core.Type{Float16}; context=context())

Creates an f16 type in the given context. The type is owned by the context.

MLIR.IR.Type — Method

Type(Core.Type{Float32}; context=context())

Creates an f32 type in the given context. The type is owned by the context.

MLIR.IR.Type — Method

Type(Core.Type{Float64}; context=context())

Creates a f64 type in the given context. The type is owned by the context.

MLIR.IR.Type — Method

Type(::Core.Type{Nothing}; context=context())

Creates a None type in the given context. The type is owned by the context.

MLIR.IR.Type — Method

Type(elements; context=context())
Type(::Core.Type{<:Tuple{T...}}; context=context())

Creates a tuple type that consists of the given list of elemental types. The type is owned by the context.

MLIR.IR.Type — Method

Type(Complex{T}) where {T}

Creates a complex type with the given element type in the same context as the element type. The type is owned by the context.

Base.:== — Method

==(t1, t2)

Checks if two types are equal.

Base.ndims — Method

ndims(type)

Returns the rank of the given ranked shaped type.

Base.parse — Method

parse(type; context=context())

Parses a type. The type is owned by the context.

Base.size — Method

size(type, i)

Returns the i-th dimension of the given ranked shaped type.

MLIR.IR.BFloat16Type — Method

BFloat16Type(; context=context())

Creates a bf16 type in the given context. The type is owned by the context.

MLIR.IR.Float8E4M3FN — Method

Float8E4M3FN(; context=context())

Creates an f8E4M3FN type in the given context. The type is owned by the context.

MLIR.IR.Float8E5M2 — Method

Float8E5M2(; context=context())

Creates an f8E5M2 type in the given context. The type is owned by the context.

MLIR.IR.FunctionType — Method

FunctionType(inputs, results; context=context())

Creates a function type, mapping a list of input types to result types.

MLIR.IR.IndexType — Method

IndexType(; context=context())

Creates an index type in the given context. The type is owned by the context.

MLIR.IR.MemRefType — Method

MemRefType(elementType, rank, shape, layout, memorySpace; location=Location(), check=false)

Creates a MemRef type with the given rank and shape, a potentially empty list of affine layout maps, the given memory space and element type, in the same context as element type. The type is owned by the context. If check=true, emits appropriate diagnostics on illegal arguments.

MLIR.IR.MemRefType — Method

MemRefType(elementType, rank, shape, memorySpace; location=Location(), check=false)

Creates a MemRef type with the given rank, shape, memory space and element type in the same context as the element type. The type has no affine maps, i.e. represents a default row-major contiguous memref. The type is owned by the context. If check=true, emits appropriate diagnostics on illegal arguments.

MLIR.IR.MemRefType — Method

MemRefType(elementType, memorySpace)

Creates an Unranked MemRef type with the given element type and in the given memory space. The type is owned by the context of element type. If check=true, emits appropriate diagnostics on illegal arguments.

MLIR.IR.OpaqueType — Method

OpaqueType(dialectNamespace, typeData; context=context())

Creates an opaque type in the given context associated with the dialect identified by its namespace. The type contains opaque byte data of the specified length (data need not be null-terminated).

MLIR.IR.TensorType — Function

TensorType(shape, elementType, encoding=Attribute(); location=Location(), check=false)

Creates a tensor type of a fixed rank with the given shape, element type, and optional encoding in the same context as the element type. The type is owned by the context. Tensor types without any specific encoding field should assign mlirAttributeGetNull to this parameter. If check=true, emits appropriate diagnostics on illegal arguments.

MLIR.IR.TensorType — Method

TensorType(elementType)

Creates an unranked tensor type with the given element type in the same context as the element type. The type is owned by the context. If check=true, emits appropriate diagnostics on illegal arguments.

MLIR.IR.VectorType — Method

VectorType(rank, shape, elementType; location=Location(), check=false)

Creates a vector type of the shape identified by its rank and dimensions, with the given element type in the same context as the element type. The type is owned by the context. If check=true, emits appropriate diagnostics on illegal arguments.

MLIR.IR.affinemap — Method

affinemap(type)

Returns the affine map of the given MemRef type.

MLIR.IR.bitwidth — Method

bitwidth(type)

Returns the bitwidth of an integer type.

MLIR.IR.context — Method

context(type)

Gets the context that a type was created with.

MLIR.IR.data — Method

mlirOpaqueTypeGetData(type)

Returns the raw data as a string reference. The data remains live as long as the context in which the type lives.

MLIR.IR.dynsize — Method

dynsize()

Returns the value indicating a dynamic size in a shaped type. Prefer isdynsize to direct comparisons with this value.

MLIR.IR.dynstrideoroffset — Method

mlirShapedTypeGetDynamicStrideOrOffset()

Returns the value indicating a dynamic stride or offset in a shaped type. Prefer isdynstrideoroffset to direct comparisons with this value.

MLIR.IR.encoding — Method

encoding(type)

Gets the 'encoding' attribute from the ranked tensor type, returning a nothing if none.

MLIR.IR.hasrank — Method

hasrank(type)

Checks whether the given shaped type is ranked.

MLIR.IR.hasstaticshape — Method

hasstaticshape(type)

Checks whether the given shaped type has a static shape.

MLIR.IR.input — Method

input(type, i)

Returns the i-th input type.

MLIR.IR.isbf16 — Method

isbf16(type)

Checks whether the given type is a bf16 type.

MLIR.IR.iscomplex — Method

iscomplex(type)

Checks whether the given type is a Complex type.

MLIR.IR.isdyndim — Method

isdyndim(type, i)

Checks wither the i-th dimension of the given shaped type is dynamic.

MLIR.IR.isdynsize — Method

isdynsize(size)

Checks whether the given value is used as a placeholder for dynamic sizes in shaped types.

MLIR.IR.isdynstrideoroffset — Method

mlirShapedTypeIsDynamicStrideOrOffset(val)

Checks whether the given value is used as a placeholder for dynamic strides and offsets in shaped types.

MLIR.IR.isf16 — Method

isf16(type)

Checks whether the given type is an f16 type.

MLIR.IR.isf32 — Method

isf32(type)

Checks whether the given type is an f32 type.

MLIR.IR.isf64 — Method

isf64(type)

Checks whether the given type is an f64 type.

MLIR.IR.isf8e4m3fn — Method

isf8e4m3fn(type)

Checks whether the given type is an f8E4M3FN type.

MLIR.IR.isf8e5m2 — Method

isf8e5m2(type)

Checks whether the given type is an f8E5M2 type.

MLIR.IR.isfunction — Method

isfunction(type)

Checks whether the given type is a function type.

MLIR.IR.isindex — Method

isindex(type)

Checks whether the given type is an index type.

MLIR.IR.isinteger — Method

isinteger(type)

Checks whether the given type is an integer type.

MLIR.IR.ismemref — Method

ismemref(type)

Checks whether the given type is a MemRef type.

MLIR.IR.isnone — Method

mlirTypeIsANone(type)

Checks whether the given type is a None type.

MLIR.IR.isopaque — Method

isopaque(type)

Checks whether the given type is an opaque type.

MLIR.IR.isrankedtensor — Method

isrankedtensor(type)

Checks whether the given type is a ranked tensor type.

MLIR.IR.isshaped — Method

isshaped(type)

Checks whether the given type is a Shaped type.

MLIR.IR.issigned — Method

issigned(type)

Checks whether the given integer type is signed.

MLIR.IR.issignless — Method

issignless(type)

Checks whether the given integer type is signless.

MLIR.IR.istensor — Method

istensor(type)

Checks whether the given type is a Tensor type.

MLIR.IR.istuple — Method

istuple(type)

Checks whether the given type is a tuple type.

MLIR.IR.isunrankedmemref — Method

mlirTypeIsAUnrankedMemRef(type)

Checks whether the given type is an UnrankedMemRef type.

MLIR.IR.isunrankedtensor — Method

isunrankedtensor(type)

Checks whether the given type is an unranked tensor type.

MLIR.IR.isunsigned — Method

isunsigned(type)

Checks whether the given integer type is unsigned.

MLIR.IR.isvector — Method

isvector(type)

Checks whether the given type is a Vector type.

MLIR.IR.layout — Method

layout(type)

Returns the layout of the given MemRef type.

MLIR.IR.memspace — Method

mlirMemRefTypeGetMemorySpace(type)

Returns the memory space of the given MemRef type.

MLIR.IR.namespace — Method

mlirOpaqueTypeGetDialectNamespace(type)

Returns the namespace of the dialect with which the given opaque type is associated. The namespace string is owned by the context.

MLIR.IR.ninputs — Method

ninputs(type)

Returns the number of input types.

MLIR.IR.nresults — Method

nresults(type)

Returns the number of result types.

MLIR.IR.result — Function

result(type, i)

Returns the i-th result type.

MLIR.IR.typeid — Method

typeid(type)

Gets the type ID of the type.